requirements.txt
flask
greenlet
gevent
#redis
python-dotenv
marshmallow
requests